feat(infra): suppression des roles associés à une délégation lors de la suppression de celle-ci

Related work items: #312109

+++ 01_account/src/Account.Core/Services/DelegationService.cs
@@ -42,6 +42,11 @@ public class DelegationService : IDelegationService
 
     public async Task DeleteDelegationAsync(int delegationId)
     {
+        if (delegationId <= 0)
+        {
+            throw new BadRequestException(Errors.BadRequestDeleteDelegationCode, Errors.BadRequestDeleteDelegationMessage);
+        }
+
         await _delegationRepository.DeleteDelegationAsync(delegationId);
     }
